remove usage of keys
