+++ src/program/protocols/cardinal/lock.rs
@@ -104,7 +104,6 @@ impl ProtocolHandler for LockProtocol {
         tx_status: TransactionStatus,
         _context: String,
         _program_context: &ProgramContext,
-        _participant_keys: Vec<&ParticipantKeys>,
     ) -> Result<(), BitVMXError> {
         let name = self.get_transaction_name_by_id(tx_id)?;
         if tx_status.confirmations == 1 {
