+++ src/program/protocols/cardinal/transfer.rs
@@ -132,7 +132,6 @@ impl ProtocolHandler for TransferProtocol {
         _tx_status: TransactionStatus,
         _context: String,
         _program_context: &ProgramContext,
-        _participant_keys: Vec<&ParticipantKeys>,
     ) -> Result<(), BitVMXError> {
         Ok(())
     }
