+++ src/program/protocols/union/accept_pegin.rs
@@ -338,7 +338,6 @@ impl ProtocolHandler for AcceptPegInProtocol {
         tx_status: TransactionStatus,
         _context: String,
         context: &ProgramContext,
-        _participant_keys: Vec<&ParticipantKeys>,
     ) -> Result<(), BitVMXError> {
         let tx_name = self.get_transaction_name_by_id(tx_id)?;
         info!(
