+++ src/program/protocols/union/advance_funds.rs
@@ -214,7 +214,6 @@ impl ProtocolHandler for AdvanceFundsProtocol {
         tx_status: TransactionStatus,
         _context: String,
         context: &ProgramContext,
-        _participant_keys: Vec<&ParticipantKeys>,
     ) -> Result<(), BitVMXError> {
         let tx_name = self.get_transaction_name_by_id(tx_id)?;
         info!(
