+++ src/program/protocols/union/full_penalization.rs
@@ -151,7 +151,6 @@ impl ProtocolHandler for FullPenalizationProtocol {
         tx_status: TransactionStatus,
         _context: String,
         program_context: &ProgramContext,
-        _participant_keys: Vec<&ParticipantKeys>,
     ) -> Result<(), BitVMXError> {
         let tx_name = self.get_transaction_name_by_id(tx_id)?;
         info!(
