+++ src/program/protocols/union/reject_pegin.rs
@@ -146,7 +146,6 @@ impl ProtocolHandler for RejectPegInProtocol {
         tx_status: TransactionStatus,
         _context: String,
         _program_context: &ProgramContext,
-        _participant_keys: Vec<&ParticipantKeys>,
     ) -> Result<(), BitVMXError> {
         let tx_name = self.get_transaction_name_by_id(tx_id)?;
         info!(
