+++ src/program/protocols/union/user_take.rs
@@ -194,7 +194,6 @@ impl ProtocolHandler for UserTakeProtocol {
         tx_status: TransactionStatus,
         _context: String,
         _program_context: &ProgramContext,
-        _participant_keys: Vec<&ParticipantKeys>,
     ) -> Result<(), BitVMXError> {
         let tx_name = self.get_transaction_name_by_id(tx_id)?;
         info!(
